+++ CHANGELOG.md
@@ -51,6 +51,7 @@
   `STATIC_DEPLOY_ESCAPE_EXCEPTIONS` is true.
 - Fix a security issue where the nonce was not verified
   for some actions until after the action had been performed.
+- Cache option values. BLOB values are not cached.
 
 ## 9.3.2 (2025-07-29)